Move season statistic helpers to helper module
